Convert enums to enum classes

+++ texture_pack.cpp
@@ -5,7 +5,7 @@
 #include "texture.h"
 
 TexturePack::TexturePack() {
-  for (int i = 0; i < TextureName::TOTAL_NUM_TEXTURES; ++i) {
+  for (int i = 0; i < static_cast<int>(TextureName::TOTAL_NUM_TEXTURES); ++i) {
     textures_.emplace_back();
   }
 }
@@ -17,11 +17,11 @@ TexturePack::~TexturePack() {
 }
 
 Texture* TexturePack::GetTexture(TextureName name) {
-  return &textures_[name];
+  return &textures_[static_cast<int>(name)];
 }
 
 void TexturePack::InitTexture(TextureName name, SDL_Renderer* renderer, TTF_Font* font) {
-  textures_[name].Init(renderer, font);
+  textures_[static_cast<int>(name)].Init(renderer, font);
 }
 
 std::string TexturePack::TexturePath(TextureName name) {
